Move `Storage` enum to `models`
