+++ lib/src/clients/web-worker-client.ts
@@ -61,7 +61,6 @@ import {
     SIGN_OUT,
     SILENT_SIGN_IN_STATE,
     START_AUTO_REFRESH_TOKEN,
-    Storage,
     UPDATE_CONFIG
 } from "../constants";
 import { AuthenticationHelper, SPAHelper, SessionManagementHelper } from "../helpers";
@@ -78,6 +77,7 @@ import {
     WebWorkerClientInterface
 } from "../models";
 import { SPACustomGrantConfig } from "../models/request-custom-grant";
+import { Storage } from "../models/storage";
 import { LocalStore, MemoryStore, SessionStore } from "../stores";
 import { SPAUtils } from "../utils";
 import { SPACryptoUtils } from "../utils/crypto-utils";
