+++ lib/src/models/storage.ts
@@ -1,0 +1,39 @@
+/**
+ * Copyright (c) 2025, WSO2 LLC. (https://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+/**
+ * Enum for the different storage types.
+ */
+export enum Storage {
+  /**
+   * Stores the session information in the local storage
+   */
+  LocalStorage = "localStorage",
+  /**
+   * Store the session information in the session storage.
+   */
+  SessionStorage = "sessionStorage",
+  /**
+   * Store the session information in the web worker.
+   */
+  WebWorker = "webWorker",
+  /**
+   * Store the session information in the browser memory.
+   */
+  BrowserMemory = "browserMemory"
+}
